% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayTable.R
\name{displayTable}
\alias{displayTable}
\title{function for displaying table}
\usage{
displayTable(data, scroll = TRUE, limit = 20)
}
\arguments{
\item{data}{Data frame. The dataframe to be displayed in table format.}

\item{scroll}{Logical. A value to have a scroll or not in the table. 
Default is TRUE.}

\item{limit}{Numeric. A value to indicate how many rows to display.
Default is 20.}
}
\value{
A table with proper formatting for html notebook
}
\description{
This is the main function for displaying data in table format
}
\examples{
data <- datasets::EuStockMarkets
dataset <- as.data.frame(data)
displayTable(dataset)
}
\author{
Vishwavani <vishwavani@mu-sigma.com>
}
\keyword{Table_Formatting}
