\name{hvt-package}
\alias{hvt-package}
\alias{hvt}
\docType{package}
\title{
  Topologically aware inference as an alternative to kernel density estimation
}
\description{
This package is modeled after the kernel density estimator of the package hypervolume. Here, the main function hypervolume_t takes in a dataframe in as many dimensions as columns, and as many samples as rows. Given two distance paramaters, it creates a sample (usually) larger than the input sample, filling in any small voids, but preserving larger voids, considered indicative of the topology of the underlying space.
  of about one to five lines is recommended.
}
\details{
TODO
}
\author{
Jānis Lazovskis

Maintainer: Janis Lazovskis <janis.lazovskis_1@rtu.lv>
}
\references{
This package is modeled after the kernel density estimator of the package hypervolume.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
