% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationResults.R
\name{OptimizationResults}
\alias{OptimizationResults}
\title{Table display of all combinations performed in `HVTMSMoptimization`}
\usage{
OptimizationResults(msm_results, show_top_global = NULL)
}
\arguments{
\item{msm_results}{Dataframe. The `all_results` object of `HVTMSMoptimization` 
output based on a specific mae_metric}

\item{show_top_global}{Numeric. To tabulate top n lowest MAE results}
}
\value{
Table of all iterations will be displayed
}
\description{
This function helps to tabulate the `HVTMSMoptimization` all iterations
performed (successful and failed combinations)
}
\author{
Vishwavani <vishwavani@mu-sigma.com>, Nithya <nithya.sn@mu-sigma.com>
}
\keyword{Hyperparameter_Tuning}
