% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMsmKN.R
\name{plotMsmKN}
\alias{plotMsmKN}
\title{plotting MAE across Number of cells}
\usage{
plotMsmKN(optimization_results)
}
\arguments{
\item{optimization_results}{List. The output of `HVTMSMoptimization` function 
specific for a mae_metric}
}
\value{
List containing visualization results for the selected MAE metric:
\item{[[plot]] }{A Plotly object of cell-level lowest MAE combination} 
\item{[[table]] }{A table of plotted data}
}
\description{
This function helps to visualize the `HVTMSMoptimization` function's
output. It creates a plotly object where the lowest MAE combination achieved for each cell 
will be plotted
}
\author{
Vishwavani <vishwavani@mu-sigma.com>, Nithya <nithya.sn@mu-sigma.com>
}
\keyword{Hyperparameter_Tuning}
